Fixed chart loading.
The issue is that the call  google.charts.load('45',
{'packages':['corechart']}) actually takes a moment to load the chart
library. This may well take a second. If, in this time, it is tried to
access the library, there will be an error, of course. Because of this,
the currently existing JavaScript chart drawing functions for piechart
and barchart have been wrapped as a callback to
google.charts.setOnLoadCallback() which waits until the library has been
loaded. Also, for loading during a Tapestry Ajax call with zones, the
<trigger> element has been added to the templates of both charts. We
need that so we can issue an event at the server side which then, in
turn, adds the JavaScript to draw the charts.
Also, the library loading call has been moved into a RequireJS module
for the index page. I couldn't get it to work from within a widget
directly.