Refined large widget view mode.
Now, when a widget is set to the large view mode, a background div with
white background is lain under the large widget to hide visible widget
parts in the background. Also, the body is set to "overflow hidden"
which avoids the scrolling of the body while working with the large
widget. Large widgets can now be scrolled. And their body is now
extended to the whole page, showing the google charts much better.

Fixes #31, #29.