Fixed a bug where not all arguments would have their preferred name set.
The issue was that the EventPostProcessingService built a Set<Argument>
of all event arguments and gave each member of this set its preferred
name. However, each event has its very own set of Argument instances.
Assume that two events share one argument with name 'AHR'. Then, both
events will have their own Argument object representing 'AHR', a and a'.
When a and a' are put into a set, only one of both will actually
contained in the set eventually. Thus, only a or a' gets its preferred
name, the other wouldn't.