Overworking the chart data distribution.
Idea: Data shouldn't go multiple times to the client. Also, filtering
often does not require contact to the server. So let the data JS
module cache data and let the client to filtering on its own.
For this, the widget layout now can switch off its default ajax
content loading. Then, the specific widgets have to take care of their
DOM elements and data loading on their own.
The data module now has methods to get data from the server, through
the index page. The idea is that all required data formats have a name
such as "relation counts" or "neighboursofnode" and such stuff.
This can then be cached within the browser.
The current issue is that the cache has to be emptied when a new
search is issued.