Fixed a bug where after search the widgets couldn't be resized.
The reason was that the WidgetManager didn't create a new
Widget object if its internal map already knew a Widget by the given
name. But with the Ajax loading the map wouldn't be refreshed but
the HTML code would. So the elements manipulated by the old Widget
object did not exist any more and the newly added Widget object
was never created because the map already contained a Widget
for each widget name.