Changed the gene name to aggregate resolution cypher query to be more robust against wrong orthology mappings.
In the data sent to Anna Lena there was a case where CIRBP was in the result list. This is because in the input list, there was Rbm3. One Rbm3 gene is somehow part of the CIRBP cluster in the current data (https://www.ncbi.nlm.nih.gov/gene/1153/ortholog/?scope=117570). And since we did infer the aggregates from their elements, the whole CIRBP cluster was added to the results.
What we now do is to directly match the input gene names to the preferred name (lowercased) field of the orthology aggregates in the database. New data with this query was sent to Anne Lena and hopefully she like it better now.