First approach to adding abbreviation full forms to the abbreviation token.
We just use the feature path `Abbreviation:/textReference/coveredText()` and apply the `textTokensFilter` used for all the running text. This should add all the tokens of the long form with an position increment of 0 to the abbreviation token. The drawback of this approach is that the tokens of the longform are effectively converted into a bag-of-words because the internal word order gets lost - all token positions are the same. If we wanted to remedy this, we would have to index a new field where the abbreviations have actually been replaced by their longforms. I figure that would break highlighting, however. So let's see if the current approach is not already good enough.