Switching to bootstrap 5.
Added a handle to the input column for getting it back after a search apart from the button in the navigation bar, which was always meant as a temporary solution. Right now, the whole column and the handle disappear to the left. There is a hover area that brings the handle back and a click on the handle brings the whole column back.
While this basically works, there are issues with the hover area because it obscures the underlying elements. I can't seem to find an elegant solution for this. So the hover area will be removed and the handle will always be visible. For this to look nice we will need some margin in the output column for the handler to have some space.