Worked on the sankey diagram to understand it better to finally be able to enlarge or shrink it when the widget is enlarged or minimized.
Changes base on the official examples of the D3 sankey plugin at https://observablehq.com/@d3/sankey-diagram. Removed quite some "g" elements that didn't seem to do much useful. Positioning now works a bit different. That's actually all not much of a deal but helps me to understand the code and adapt it.