Beginning to add the 'tab' persistence strategy.
Which just means that we can have multiple search sessions which are separate from each other. Additionally, we'd like to realize a kind of "sub-session" which is created when we clicked onto something in a chart and then get more detailed data. The context of the detailed view would be the original search.