Create specific fields for argument1 and argument2 (#132).
Until now we just had the "arguments" arrays for ids, texts and so on. However, this poses an issue when the same gene appears in the A list and in the B list. This causes the retrieval of all relations that have that gene as any one event partner, no matter the other event partner since both, A and B query, are fulfilled. Such cases could be removed after retrieval but that would mean that a lot of unrequired events are being retrieved just to filter them out.
Instead, the index now has specific fields for both arguments.
Note that the code not yet makes use of those fields.