Activate the delete-before-index function in the ES consumer.
It seems that the relation IDs are not stable. After each indexing, there were a few more documents in the index. This would also a problem when updating documents that have actually different entities or relations. By deleting all index document associated with an input document first, this issue is circumvented.