Cut down on superfluous ES queries.
We had one full query, the legacy one that received all results with highlighting. This was used for the charts. But they don't require the highlighting. The table widget uses paging and did ignore the existing result at all.
This has been changed now in a way where the initial result is the first page for the table, with highlighting. This is now reused so that we only issue a single query to ES right now.
However, this also means that the charts are currently restricted to the first result page which is of course not what we want. Thus, we need to issue a second search for the charts that must be as efficient as possible. For this we also once introduced the Neo4j query. But that was slow. Perhaps we can revisit this later in the hope of a large speedup.