Open a bootstrap modal when downloading the data.
It should explain that downloading can take a while. The goal is to now let the model trigger the actual download and only allow closing the modal after the download is done to avoid further interaction with the web page while the data is assembled. Possibly, a cancel button would also be interesting if this can work on a simple way.