Fix #159.
The issue was that the paged result was used in TableResultWidget. Just using the "unrolled" result did not work, however, since the default unrolled result is meant for the charts and does not carry contextual information. Thus, there is now unrolled data for charts and unrolled data for download. The latter one will be rather large due to all the text. It is still cached in GepiData but with a WeakReference so the GC can evict it, if necessary.