Remove unused Sankey control options.
Some options just didn't do anything really. For now, the Sankey diagrams look fine as they are.
Also added a tooltip to the 'Show "other" node' button and actually had to dispose of the tooltip manually when clicking the button because after `redraw()` is called, the tooltip wouldn't disappear anymore. Most like because at the end of `redraw()` we also call `initTooltips()`.