Polish chart widgets.
Add tooltips, better layout, small stuff. Contained in this commit is also some - now outcommented - code for the configuration (add / show / hide / remove) columns from the TableResultWidget. Not quite working right, when columns are removed, so are the checkboxes. The checkbox layout was also awkward so I commented this out for the time being. It's not like anyone had requested a possibility to how or hide columns.