Exclude tid-preferredName mappings from the tid-to-homology-aggregate-name mapping.
Because that should actually used for the homology - are more precisely orthology - relationship between genes and their gene_group and resulting othology aggregate. We already have the tid2concept mapping file that can be used in the filter chain just after the mapping from tid to orthology aggregate. When a tid does not have such a mapping, it can then be converted to the concept preferred name via the tid2concept mapping.