Do not load geneInfo in EventResponseProcessingService / make difference between load all for charts and load all for Excel sheet creation.
We loaded the geneInfo in the EventResponseProcessingService in part to handle the FamPlex and HGNC Group concept IDs because we needed their original IDs (and labels). Instead, we now add the original IDs directly into the index for FamPlex and HGNC Group concepts, too (just as with the genes).
Also, the event retrieval code handled two cases well, both for the Website (charts and table widget) but did neglect the requirements for the complete download for Excel sheet creation. This has been fixed.