Create a separate mapping for FPLX/HGNCG to their aggregates to fix the more-than-two-concept-Ids-issue.
Through the concatenation of FPLX/HGNCG IDs in the GepiFamplexIdAssigner, there are two IDs on one entity because it has an FPLX and also and HGNC Group ID. We solve this by mapping to their equal-name aggregate. But we want to do this only for FPLX/HGNCG and not for normal gene mentions.